#74 rename: ShopImgInfo.tsx 파일 컴포넌트 쪼갬
